Extract out common transform functions to `transform_utils` file
Sine these are platform agnostic transforms, we can use these for iOS as well instead of recreating them
